+++ src/Chord.ts
@@ -194,7 +194,7 @@ export default class Chord implements IPitchContainer {
     }
 
     /**
-     * Adds one or more pitches to a chord, based on chord degrees. For example, addDegrees([3,5]) would add a 3rd & 5th to the chord. The degrees added will always default to the major/perfect variety, though if the scale parameter is used, then the minor/diminished/augmented versions can be used if they would be deemed to be a better fit within the scale.
+     * Adds one or more pitches to a chord, based on chord degrees. For example, addDegrees([3,5]) would add a 3rd & 5th to the chord. The degrees added will always default to the major/perfect variety, though if the scale parameter is used, then the minor/diminished/augmented versions can be used if they would be deemed to be a better fit within the scale. If you want to force the use of the minor/diminished version of the degree, then the degree can be provided as a negative, for example: addDegrees([-3,5])
      * @param degrees The degrees of the chord to be added.
      * @param scale Optional, used to allow for better fitting pitch selection.
      * @returns Returns the chord instance, so that method calls can be chained together.
@@ -203,8 +203,16 @@ export default class Chord implements IPitchContainer {
         if (this.root == null)
             throw new Error('You cannot add degrees to the chord before its root has been set');
         
-        //Map each degree to one less, this makes them 0-based and far easier to work with
-        for (const degree of degrees.map(x => x - 1)) {
+        for (let degree of degrees) {
+            let forceFallback = false;
+            if (degree < 0) {
+                forceFallback = true;
+                degree = Math.abs(degree);
+            }
+
+            //Map each degree to one less, this makes them 0-based and far easier to work with
+            degree--;
+
             let pitch = this.root;
             if (degree >= 7)
                 pitch += Math.floor(degree / 7) * 12;
@@ -230,7 +238,10 @@ export default class Chord implements IPitchContainer {
                     degreeAddition = { main: 11, fallback: 10 };
                     break;
             }
-            if (
+
+            if (forceFallback)
+                pitch += degreeAddition.fallback
+            else if (
                 scale == null ||
                 scale.contains(pitch + degreeAddition.main) ||
                 !scale.contains(pitch + degreeAddition.fallback)
@@ -244,42 +255,74 @@ export default class Chord implements IPitchContainer {
         return this;
     }
 
-    /** Adds a 2nd degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 2nd degree to the chord, see addDegrees for more info. 
+     * */
     add2nd(scale: Scale = null): Chord {
         return this.addDegrees([2], scale);
     }
 
-    /** Adds a 3rd degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 3rd degree to the chord, see addDegrees for more info. 
+     * */
     add3rd(scale: Scale = null): Chord {
         return this.addDegrees([3], scale);
     }
 
-    /** Adds a 4th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 4th degree to the chord, see addDegrees for more info. 
+     * */
     add4th(scale: Scale = null): Chord {
         return this.addDegrees([4], scale);
     }
 
-    /** Adds a 5th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 5th degree to the chord, see addDegrees for more info. 
+     * */
     add5th(scale: Scale = null): Chord {
         return this.addDegrees([5], scale);
     }
 
-    /** Adds a 6th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 6th degree to the chord, see addDegrees for more info. 
+     * */
     add6th(scale: Scale = null): Chord {
         return this.addDegrees([6], scale);
     }
 
-    /** Adds a 7th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 7th degree to the chord, see addDegrees for more info. 
+     * */
     add7th(scale: Scale = null): Chord {
         return this.addDegrees([7], scale);
     }
 
-    /** Adds a 9th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 9th degree to the chord, see addDegrees for more info. 
+    */
     add9th(scale: Scale = null): Chord {
         return this.addDegrees([9], scale);
     }
 
-    /** Adds a 11th degree to the chord, see addDegrees for more info. */
+    /** 
+     * @deprecated Just use addPitches instead
+     * 
+     * Adds a 11th degree to the chord, see addDegrees for more info. 
+     * */
     add11th(scale: Scale = null): Chord {
         return this.addDegrees([11], scale);
     }